BACKPORT: Clear icons for app/ime keyboard shortcuts in shortcut helper.
If application supplied keyboard shortcuts have an icon, clear that icon
before processing it in the shortcut helper. Icons are not intended for
use by applications.

Additionally resolves a race condition between receiving IME and app
specific shortcuts, and closes the shortcut helper if it is showing
while a user switch takes place.

Bug: 331180422
Test: See b/331180422; atest KeyboardShortcutsTest KeyboardShortcutListSearchTest

Change-Id: Ia628886e80e956d0c423d7a7ebe67b2b8ca8c264
Signed-off-by: Alexey Shelby <shelbyhell@proton.me>